# src/main/resources/swagger.yaml
...
swagger: "2.0"
info:
  description: "Medikm Apis Description"
  version: "1.0"
  title: "Medikm Apis"
...
schemes:
- "http"
- "https"
paths:
  /case/addCase:
    post:
      tags:
      - "Case Controller"
      summary: "Add Case"
      description: "Case Added."
      operationId: "createCase"
      parameters:
      - in: "body"
        name: "body"
        description: "Case Request"
        required: false
        schema:
          $ref: "#/definitions/Case Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Case Response"
        500:
          description: "Failed | Disease Type does not found"
          schema:
            $ref: "#/definitions/Case Response"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/Case Response"
  /case/editCase:
    post:
      tags:
      - "Case Controller"
      summary: "Edit Case"
      description: "Case Edited."
      operationId: "editCase"
      parameters:
      - in: "body"
        name: "body"
        description: "Case Request"
        required: false
        schema:
          $ref: "#/definitions/Case Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed | Disease Type does not found"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /case/getSimilarCases:
    post:
      tags:
      - "Case Controller"
      summary: "Get All Similar Case"
      description: "Get All Similar Case."
      operationId: "getAllSimilarCase"
      parameters:
      - in: "body"
        name: "body"
        description: "Similar Case Request"
        required: false
        schema:
          $ref: "#/definitions/Similar Case Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Case Response"
        500:
          description: "Failed | Disease Type does not found"
          schema:
            $ref: "#/definitions/Case Response"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/Case Response"
  /episode/addEpisode:
    post:
      tags:
      - "Episode Controller"
      summary: "Add Episode"
      description: "Episode Added."
      operationId: "addEpisode"
      parameters:
      - in: "body"
        name: "body"
        description: "Episode Request"
        required: false
        schema:
          $ref: "#/definitions/Episode Response"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Episode Response"
        500:
          description: "Failed | No disease type found"
          schema:
            $ref: "#/definitions/Episode Response"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/Episode Response"
  /episode/editEpisode:
    post:
      tags:
      - "Episode Controller"
      summary: "Edit Episode"
      description: "Episode Edited."
      operationId: "editEpisode"
      parameters:
      - in: "body"
        name: "body"
        description: "Episode Request"
        required: false
        schema:
          $ref: "#/definitions/Episode Response"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Episode Response"
        500:
          description: "Failed | No disease type found"
          schema:
            $ref: "#/definitions/Episode Response"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/Episode Response"
...
definitions:
  Biopsy:
    type: "object"
    properties:
      lesionLocation:
        type: "string"
      needleGauge:
        type: "integer"
        format: "int32"
      postBiopsyRecommendedMRI:
        type: "integer"
        format: "int32"
      postBiopsyRelatedMRIControlateralBreast:
        type: "integer"
        format: "int32"
      lesionCompletelyRemoved:
        type: "integer"
        format: "int32"
      complications:
        type: "string"
      timeBiopsySampleTaken:
        type: "string"
        format: "date-time"
      timeBiopsySampleSent:
        type: "string"
        format: "date-time"
      biopsyMargins:
        type: "integer"
        format: "int32"
      biopsyMarginsComm:
        type: "string"
      lesionType:
        $ref: "#/definitions/LesionType"
      biopsyTypeId:
        $ref: "#/definitions/BiopsyType"
      caseId:
        $ref: "#/definitions/Case1"
      encounter:
        $ref: "#/definitions/Encounter"
      sninjMethod:
        type: "string"
...
  CcbrQuestion:
    type: "object"
    properties:
      questionId:
        type: "integer"
        format: "int32"
      exclude1:
        type: "boolean"
        default: false
      exclude2:
        type: "boolean"
        default: false
      exclude3:
        type: "boolean"
        default: false
      questiomtext:
        type: "string"
      rational:
        type: "string"
      referenceKnowledge:
        type: "integer"
        format: "int32"
      solutionAvailable:
        type: "integer"
        format: "int32"
      source:
        type: "string"
      comments:
        type: "string"
      ccbrSolutionCollection:
        type: "array"
        items:
          $ref: "#/definitions/CcbrSolution"
      category:
        $ref: "#/definitions/SolCategory"
      keyConcept1:
        $ref: "#/definitions/KeyConcept"
      keyConcept3:
        $ref: "#/definitions/KeyConcept"
      keyConcept2:
        $ref: "#/definitions/KeyConcept"
      tumorBoardDiscussionCollection:
        type: "array"
        items:
          $ref: "#/definitions/TumorBoardDiscussion"
      tbDiscussionForumQuestionCollection:
        type: "array"
        items:
          $ref: "#/definitions/TbDiscussionForumQuestion"
      qrpage:
        type: "boolean"
        default: false
      qrper:
        type: "boolean"
        default: false
      qrppr:
        type: "boolean"
        default: false
      qrpgenetics:
        type: "boolean"
        default: false
      qrpgenomics:
        type: "boolean"
        default: false
      qrpcancerType:
        type: "boolean"
        default: false
      qrptumorSize:
        type: "boolean"
        default: false
      qrpnodes:
        type: "boolean"
        default: false
      qrpworkingTNM:
        type: "boolean"
        default: false
      qrpfinalTNM:
        type: "boolean"
        default: false
      qrpgrade:
        type: "boolean"
        default: false
      qrpher2neu:
        type: "boolean"
        default: false
      qrpher2neuSISH:
        type: "boolean"
        default: false
      qrpdistantMet:
        type: "boolean"
        default: false
      qrpgender:
        type: "boolean"
        default: false
      qrpmenopause:
        type: "boolean"
        default: false
      qrpsignificanceRisk:
        type: "boolean"
        default: false
      qrpsignificanceComorbities:
        type: "boolean"
        default: false
...
  Chemotherapy:
    type: "object"
    properties:
      chemotherapyPK:
        $ref: "#/definitions/ChemotherapyPK"
      baseLineMegaScan:
        type: "integer"
        format: "int32"
      baseLineMegaScanComm:
        type: "string"
      pretreatmentWeight:
        type: "number"
        format: "double"
      postTreatmentWeight:
        type: "number"
        format: "double"
      startDtCtpy:
        type: "string"
        format: "date-time"
      endDtCtpy:
        type: "string"
        format: "date-time"
      ctpyRoute:
        type: "number"
        format: "double"
      ctpDose:
        type: "number"
        format: "double"
      doseReductionNeeded:
        type: "integer"
        format: "int32"
      noCyclesAdministered:
        type: "integer"
        format: "int32"
      enrolledClinicalTrial:
        type: "integer"
        format: "int32"
      eligibleClinicalTrial:
        type: "integer"
        format: "int32"
      sideEffectsExperienced:
        type: "integer"
        format: "int32"
      sideEffectsExperiencedComm:
        type: "string"
      allergicEvents:
        type: "integer"
        format: "int32"
      allergicEventsComm:
        type: "string"
      antracyclinesAdministered:
        type: "integer"
        format: "int32"
      antracyclinesAdministeredDose:
        type: "number"
        format: "double"
      hospitalizationToxicityDuringTreatment:
        type: "integer"
        format: "int32"
      neurotoxicityImpairsActivitiesDailyLiving:
        type: "integer"
        format: "int32"
      cardiotoxicity:
        type: "integer"
        format: "int32"
      cardiotoxicityComm:
        type: "string"
      postChemoMagaScan:
        type: "integer"
        format: "int32"
      postChemoMagaScanComm:
        type: "string"
      reasonStoppingCtpy:
        type: "string"
      pretreatmentPlateletCount:
        type: "integer"
        format: "int32"
      pretreatmentPlateletCountComm:
        type: "string"
      treatmentRelatedAnemia:
        type: "integer"
        format: "int32"
      treatmentRelatedNeutropenia:
        type: "integer"
        format: "int32"
      treatmentRelatedTrombocytopenia:
        type: "integer"
        format: "int32"
      treatmentRelatedLymphocytopenia:
        type: "integer"
        format: "int32"
      palliativeLine:
        type: "integer"
        format: "int32"
      chemoResponse:
        type: "string"
      radiosensitising:
        type: "integer"
        format: "int32"
      chemoStopReason:
        type: "string"
      chemoIntension:
        type: "string"
      chemoCompid:
        type: "integer"
        format: "int32"
      chemoType:
        $ref: "#/definitions/ChemoType"
      ctpyAgentId:
        $ref: "#/definitions/ChemoAgent"
      ctpyProtocolId:
        $ref: "#/definitions/ChemoProtocol"
      eligibleClinicalTrialID:
        $ref: "#/definitions/ClinicalTrial"
      caseId:
        $ref: "#/definitions/Case1"
      enrolledClinicalTrialID:
        $ref: "#/definitions/ClinicalTrial"
      episode:
        $ref: "#/definitions/Episode"
...
  DiagnosticWorkup:
    type: "object"
    properties:
      encounterId:
        type: "integer"
        format: "int32"
      dxWkupTypeId:
        $ref: "#/definitions/WorkupKindType"
      encounter:
        $ref: "#/definitions/Encounter"
      caseId:
        $ref: "#/definitions/Case1"
      dxwkupResult:
        type: "string"
      dxwkupResultDesc:
        type: "string"
      dxwkupResultReportId:
        type: "integer"
        format: "int32"
  DiseaseStatus:
    type: "object"
    properties:
      typeId:
        type: "integer"
        format: "int32"
      episodeCollection:
        type: "array"
        items:
          $ref: "#/definitions/Episode"
      dname:
        type: "string"
...
  Encounter:
    type: "object"
    properties:
      encounterId:
        type: "integer"
        format: "int32"
      encounterDt:
        type: "string"
        format: "date-time"
      facilityId:
        type: "integer"
        format: "int32"
      providerNote:
        type: "string"
      pregnancyStatus:
        type: "integer"
        format: "int32"
      monthPregnancy:
        type: "integer"
        format: "int32"
      clinicalTrialOffered:
        type: "integer"
        format: "int32"
      clinicalTrialDeclined:
        type: "integer"
        format: "int32"
      enBmi:
        type: "integer"
        format: "int32"
      encounterTypeId:
        $ref: "#/definitions/EncounterType"
      caseId:
        $ref: "#/definitions/Case1"
      episodeId:
        $ref: "#/definitions/Episode"
      surgery:
        $ref: "#/definitions/Surgery"
      diagnosticWorkup:
        $ref: "#/definitions/DiagnosticWorkup"
      encounterPhysicianCollection:
        type: "array"
        items:
          $ref: "#/definitions/EncounterPhysician"
      biopsy:
        $ref: "#/definitions/Biopsy"
      pathology:
        $ref: "#/definitions/Pathology"
      physicalExam:
        $ref: "#/definitions/PhysicalExam"
      enheight:
        type: "number"
        format: "double"
      enweight:
        type: "number"
        format: "double"
...
  EncounterPhysician:
    type: "object"
    properties:
      encounterPhysicianPK:
        $ref: "#/definitions/EncounterPhysicianPK"
      physicanEncounterNote:
        type: "string"
      physician:
        $ref: "#/definitions/Physician"
      encounter:
        $ref: "#/definitions/Encounter"
...
  EncounterType:
    type: "object"
    properties:
      typeId:
        type: "integer"
        format: "int32"
      encounterCollection:
        type: "array"
        items:
          $ref: "#/definitions/Encounter"
      ename:
        type: "string"
...
  Episode:
    type: "object"
    properties:
      episodeId:
        type: "integer"
        format: "int32"
      firstEncounterDt:
        type: "string"
        format: "date-time"
      lastEncounterDt:
        type: "string"
        format: "date-time"
      dtRemission:
        type: "string"
        format: "date-time"
      dtInitialComplaint:
        type: "string"
        format: "date-time"
      dtResolutionComplaints:
        type: "string"
        format: "date-time"
      healthProblemType:
        type: "string"
      problemUnsetDt:
        type: "string"
        format: "date-time"
      symptomsComplaintsExperienced:
        type: "string"
      sideOfCancer:
        type: "string"
      cancerLocation:
        type: "string"
      hrt:
        type: "boolean"
        default: false
      relativeIntraoperativeFindings:
        type: "string"
      imagingFindings:
        type: "string"
      biopsyFindings:
        type: "string"
      pathologyFindings:
        type: "string"
      frozenSectionFindings:
        type: "string"
      encounterCollection:
        type: "array"
        items:
          $ref: "#/definitions/Encounter"
      tumorBoardDiscussionCollection:
        type: "array"
        items:
          $ref: "#/definitions/TumorBoardDiscussion"
      tbDiscussionForumCollection:
        type: "array"
        items:
          $ref: "#/definitions/TbDiscussionForum"
      chemotherapy:
        $ref: "#/definitions/Chemotherapy"
      diseaseType:
        $ref: "#/definitions/DiseaseStatus"
      caseId:
        $ref: "#/definitions/Case1"
      episodePhysicianCollection:
        type: "array"
        items:
          $ref: "#/definitions/EpisodePhysician"
      tumorBoardDiscussionQuestionList:
        type: "array"
        items:
          $ref: "#/definitions/TumorBoardDiscussionQuestion"
      recommendedPlan:
        type: "string"
      nextVisitDate:
        type: "string"
        format: "date-time"
      epheight:
        type: "number"
        format: "double"
      epweight:
        type: "number"
        format: "double"
...
  EpisodePhysician:
    type: "object"
    properties:
      episodePhysicianPK:
        $ref: "#/definitions/EpisodePhysicianPK"
      physicanNotes:
        type: "string"
      dateOfFirstPEncounter:
        type: "string"
        format: "date-time"
      physician:
        $ref: "#/definitions/Physician"
      episode:
        $ref: "#/definitions/Episode"
...
  MedikmResponse:
    type: "object"
    properties:
      responseCode:
        type: "string"
        description: "Response Code"
      responseMessage:
        type: "string"
        description: "Response Message"
      authenticationKey:
        type: "string"
        description: "Authentication Key"
    description: "MedikmResponse class used for returning medikmresponse"
...
  PhysicalExam:
    type: "object"
    properties:
      encounterId:
        type: "integer"
        format: "int32"
      performanceScore:
        type: "number"
        format: "double"
      palpableNodes:
        type: "integer"
        format: "int32"
      skinChanges:
        type: "integer"
        format: "int32"
      nippleChanges:
        type: "integer"
        format: "int32"
      encounter:
        $ref: "#/definitions/Encounter"
      caseId:
        $ref: "#/definitions/Case1"
      cbedt:
        type: "string"
        format: "date-time"
      cbereport:
        type: "string"
      cbenormal:
        type: "integer"
        format: "int32"
...
  Surgery:
    type: "object"
    properties:
      encounterId:
        type: "integer"
        format: "int32"
      patientEligibleBcSurg:
        type: "integer"
        format: "int32"
      surgicalConsultationDt:
        type: "string"
        format: "date-time"
      breastConservationSurgeryMargins:
        type: "integer"
        format: "int32"
      resectioninvolvedMargin:
        type: "integer"
        format: "int32"
      resectioninvolvedMarginComm:
        type: "string"
      resectionEexcisionCavity:
        type: "integer"
        format: "int32"
      orientationResectionSn:
        type: "integer"
        format: "int32"
      orientationResectionSnComm:
        type: "string"
      radTherapistConsultationDt:
        type: "string"
        format: "date-time"
      skinSpearingMastectomyRTConsult:
        type: "integer"
        format: "int32"
      skinSpearingMastectomyRTConsultDate:
        type: "string"
        format: "date-time"
      sentinelNodeBiopsy:
        type: "integer"
        format: "int32"
      sentinelNodeIdentifiedDye:
        type: "integer"
        format: "int32"
      noSentineNodesRemoved:
        type: "integer"
        format: "int32"
      axillaryNodeDissection:
        type: "integer"
        format: "int32"
      axillaryDisectionNumberNode:
        type: "integer"
        format: "int32"
      ambuBCSurgUnplanONightStay:
        type: "integer"
        format: "int32"
      ambulatoryBCSurgery:
        type: "integer"
        format: "int32"
      plasticSurgeonConsult:
        type: "integer"
        format: "int32"
      plasticSurgeonConsultDate:
        type: "string"
        format: "date-time"
      reconstructiveBSImplant:
        type: "integer"
        format: "int32"
      reconstructiveBSDate:
        type: "string"
        format: "date-time"
      immediateBreastReconstruction:
        type: "integer"
        format: "int32"
      delayedBreastReconstruction:
        type: "integer"
        format: "int32"
      reconstructiveBSComplications:
        type: "integer"
        format: "int32"
      reconstructiveBSComplicationsComm:
        type: "string"
      contraindicationBC:
        type: "integer"
        format: "int32"
      contraindicationBCComm:
        type: "string"
      medConsultPreSurgery:
        type: "integer"
        format: "int32"
      medConsultPreSurgeryDt:
        type: "string"
        format: "date-time"
      marginMeasurement:
        type: "integer"
        format: "int32"
      tumorInk:
        type: "integer"
        format: "int32"
      dateOfSurgery:
        type: "string"
        format: "date-time"
      primaryProcInd:
        type: "integer"
        format: "int32"
      reasonForRetToOr:
        type: "integer"
        format: "int32"
      unexpectedRetToOr:
        type: "integer"
        format: "int32"
      postOpEvent30day:
        type: "integer"
        format: "int32"
      dispositionAfterSurg:
        type: "integer"
        format: "int32"
      surMarginAssessment:
        type: "integer"
        format: "int32"
      breastSide:
        type: "string"
      brSurgLocation:
        type: "integer"
        format: "int32"
      anesthesiaType:
        type: "string"
      surTimeStatus:
        type: "integer"
        format: "int32"
      reOperation:
        type: "integer"
        format: "int32"
      intraopPRBC:
        type: "integer"
        format: "int32"
      tumourSpill:
        type: "integer"
        format: "int32"
      plasticSurgeryType:
        $ref: "#/definitions/PlasticSurgeryType"
      surgeryTypeId:
        $ref: "#/definitions/SurgeryType"
      caseId:
        $ref: "#/definitions/Case1"
      encounter:
        $ref: "#/definitions/Encounter"
      orentryTime:
        type: "string"
        format: "date-time"
      orexitTime:
        type: "string"
        format: "date-time"
      asaclassification:
        type: "integer"
        format: "int32"
...
  TbDiscussionForum:
    type: "object"
    properties:
      discussiondt:
        type: "string"
        format: "date-time"
      primaryQuestionAskedtoTB:
        type: "string"
      pubResearchUsed:
        type: "string"
      patientInputConsidered:
        type: "integer"
        format: "int32"
      patientInputConsideredComment:
        type: "string"
      patientPreferences:
        type: "string"
      consensusReached:
        type: "integer"
        format: "int32"
      consensusReachedComment:
        type: "string"
      finalTBrecommendation:
        type: "string"
      rescheduledFUTBDiscussion:
        type: "integer"
        format: "int32"
      rescheduledFUTBDiscussionComm:
        type: "string"
      optionsGenerated:
        type: "string"
      convergingopinions:
        type: "integer"
        format: "int32"
      convergingopinionsComment:
        type: "string"
      divergingOpinios:
        type: "integer"
        format: "int32"
      divergingOpiniosComment:
        type: "string"
      additionalEvidence:
        type: "string"
      literatureConsulted:
        type: "string"
      isAlgorithmUsed:
        type: "integer"
        format: "int32"
      isAlgorithmUsedComment:
        type: "string"
      clinicalTrailsEligible:
        type: "integer"
        format: "int32"
      clinicalTrailsDiscussed:
        type: "integer"
        format: "int32"
      clinicalTrailsComment:
        type: "string"
      comparativeEffectivenessAnalysis:
        type: "integer"
        format: "int32"
      optionsConsidered:
        type: "string"
      ccbrQuestion:
        $ref: "#/definitions/CcbrQuestion"
      case1:
        $ref: "#/definitions/Case1"
      episode:
        $ref: "#/definitions/Episode"
      recommendedPlanLikeCollection:
        type: "array"
        items:
          $ref: "#/definitions/RecommendedPlanLike"
      discussionEndTime:
        type: "string"
        format: "date-time"
      status:
        type: "string"
      tbDiscussionForumQuestionList:
        type: "array"
        items:
          $ref: "#/definitions/TbDiscussionForumQuestion"
      tbdiscussionId:
        type: "integer"
        format: "int32"
      nccnguidelineMet:
        type: "integer"
        format: "int32"
      nccnguidelineMetComment:
        type: "string"
...
  TbDiscussionParticipants:
    type: "object"
    properties:
      tbDiscussionParticipantsPK:
        $ref: "#/definitions/TbDiscussionParticipantsPK"
      physicianSpeciality:
        type: "integer"
        format: "int32"
      tumorBoardDiscussion:
        $ref: "#/definitions/TumorBoardDiscussion"
...
  TumorBoardDiscussion:
    type: "object"
    properties:
      gettBDiscussionId:
        type: "integer"
        format: "int32"
      discussiondt:
        type: "string"
        format: "date-time"
      primaryQuestionAskedtoTB:
        type: "string"
      pubResearchUsed:
        type: "string"
      getnCCNGuidelineMet:
        type: "integer"
        format: "int32"
      getnCCNGuidelineMetComment:
        type: "string"
      patientInputConsidered:
        type: "integer"
        format: "int32"
      patientInputConsideredComment:
        type: "string"
      patientPreferences:
        type: "string"
      consensusReached:
        type: "integer"
        format: "int32"
      consensusReachedComment:
        type: "string"
      finalTBrecommendation:
        type: "string"
      rescheduledFUTBDiscussion:
        type: "integer"
        format: "int32"
      rescheduledFUTBDiscussionComm:
        type: "string"
      optionsGenerated:
        type: "string"
      convergingopinions:
        type: "integer"
        format: "int32"
      convergingopinionsComment:
        type: "string"
      divergingOpinios:
        type: "integer"
        format: "int32"
      divergingOpiniosComment:
        type: "string"
      additionalEvidence:
        type: "string"
      literatureConsulted:
        type: "string"
      isAlgorithmUsed:
        type: "integer"
        format: "int32"
      isAlgorithmUsedComment:
        type: "string"
      clinicalTrailsEligible:
        type: "integer"
        format: "int32"
      clinicalTrailsDiscussed:
        type: "integer"
        format: "int32"
      clinicalTrailsComment:
        type: "string"
      comparativeEffectivenessAnalysis:
        type: "integer"
        format: "int32"
      optionsConsidered:
        type: "string"
      questionId:
        $ref: "#/definitions/CcbrQuestion"
      caseId:
        $ref: "#/definitions/Case1"
      episodeId:
        $ref: "#/definitions/Episode"
      tbDiscussionForumAnswereCollection:
        type: "array"
        items:
          $ref: "#/definitions/TbDiscussionForumAnswere"
      tbDiscussionParticipantsCollection:
        type: "array"
        items:
          $ref: "#/definitions/TbDiscussionParticipants"
      tbDiscussionForumQuestionCollection:
        type: "array"
        items:
          $ref: "#/definitions/TbDiscussionForumQuestion"
      tumorBoardDiscussionQuestionList:
        type: "array"
        items:
          $ref: "#/definitions/TumorBoardDiscussionQuestion"
      discussionEndTime:
        type: "string"
        format: "date-time"
      comment:
        type: "string"
      pathway:
        type: "string"
      subPathway:
        type: "string"
      tbdiscussionId:
        type: "integer"
        format: "int32"
      nccnguidelineMet:
        type: "integer"
        format: "int32"
      nccnguidelineMetComment:
        type: "string"
...
  TumorBoardDiscussionQuestion:
    type: "object"
    properties:
      questionId:
        type: "integer"
        format: "int32"
      questionText:
        type: "string"
      dateAsked:
        type: "string"
        format: "date-time"
      episodeId:
        $ref: "#/definitions/Episode"
      caseId:
        $ref: "#/definitions/Case1"
      tbDiscussionId:
        $ref: "#/definitions/TumorBoardDiscussion"
      tumorBoardDiscussionAnswersList:
        type: "array"
        items:
          $ref: "#/definitions/TumorBoardDiscussionAnswers"
...